[MOVE] details 패키지 생성 및 MemberDetails 이동 #145
