+++ src/main/java/com/sports/server/auth/details/MemberDetails.java
@@ -1,4 +1,4 @@
-package com.sports.server.auth;
+package com.sports.server.auth.details;
 
 import com.sports.server.command.member.domain.Member;
 import java.util.ArrayList;
