+++ src/main/java/com/sports/server/auth/presentation/AuthController.java
@@ -1,7 +1,7 @@
 package com.sports.server.auth.presentation;
 
 import com.sports.server.auth.dto.JwtResponse;
-import com.sports.server.auth.MemberDetails;
+import com.sports.server.auth.details.MemberDetails;
 import com.sports.server.auth.application.AuthService;
 import com.sports.server.auth.dto.LoginVO;
 import com.sports.server.auth.utils.CookieUtil;
